INTYGFV-14651: Accordion should not open when pressing spacebar in textinput
